[msbuild] goodbye Jack & Jill.
due to: https://android-developers.googleblog.com/2017/03/future-of-java-8-language-feature.html

Q: Is it safe to just kill them right now?
A: Yes. Projects that had <UseJackAndJill>true</UseJackAndJill>, if ever any,
should still just build using Java8 javac with no functional regression.

Jack and Jill option was introduced because future (of that point) Google
was going to switch to "full Jack and Jill" world.

But in this World line everything goes back to javac as if nothing happened.