[Xamarin.Android.Build.Tasks] AAPT packaging error ignored (part 2)
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=53054

The RegEx was fixed in commit 252fa10e. However that was not
the entire problem. Firstly there was a change of the aapt
process exiting before all the output was processed. Which would
mean could miss out on messages. Secondly occasionally just
logging High priority messages with an "error" format did NOT
result in that being turned into an error.

So we need to rework the code to raise actual Errors! along with
file and line numbers. This give us a consistent experience.

This highlighted a bug in one of the unit tests specifically Bug12935.
In this test we were expecting a build to work when targeting API
15 (v4.0.3) if it used the screenOrientation value of sensorPortait.
However it turns out that the google tooling has been reporting that
usage as an AAPT error for a loooong time. This bug however was masking
it.

So this commit also fixes up that test to make sure that the build
does fail with the appropriate error message.