[Xamarin.Android.Build.Tasks] Improve Aapt output processing (#495)
The commit 049b2c8f added some ManualResetEvents to control
how we wait for the process to complete output. The problem
with that commit is that in the Output events we did not set
the events if they had completed (i.e we got null data).

This commit corrects that issue. We increased the timeout so that
event if the process is outputting allot of data we will still get
it.