Correct "Exists" logic on monodoc debug files
The previous commit's attempts to copy the pdb/mdb files for monodoc
had a problem where they would be evaluated too early (probably before
the files existed) and in the wrong directory. To address this, the
files are instead specified as a MonoDocCopyItemOptional, a new target
_GetMonodocItems is created and the ItemGroups for
_MonoDocCopyItems/_MonoDocInstalledItems are evaluated in this target
with correct Exists logic for the Optional files. Additional
DependsOnTargets attributes were distributed through the file to
ensure this new target is evaluated at the correct time.