[Xamarin.Android.Build.Tests] Move tests into a Shared Project so they can be reused upstream.
Rather than have all the msbuild tests in an assembly this commit
moves them into a shared project. This means the source code for
the tests can be easily reused upstream. It also means we can make
use of things like partial classes to change the tests depending on
which repo we are running them from. For example having tests in this
repo only run in Release mode but have both Release and Debug upstream.