[Xamarin.Android.Build.Tasks] Unable to compile projects with certain characters or spaces when AOT+LLVM is enabled
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=52644

It appears we were double double quoting the path to the assembly
which is to be AOT'd. This was causing a problem for the
tooling. This commit removes the additional quotes so
that we only quote the path once.

It also adds a unit test which has spaces (and brackets) in
the path and updates the test framework to allow to paths with
spaces.