First step towards support for Android-O Preview API.
The changes are big due to the structural changes in Android SDK, particularly
that "android" command line tool is gone and they had to be rewritten.
Samples that depended on "create" or "update" command had to be checked in
as already-created state so that they don't have to bring further dependencies.

Changes in the library (Mono.Android) itself is not very big at all.