[bundle] Include mono's debug symbols (#513)
Microsoft has (several?) [symbol servers][0], which is a repository
"somewhere" which contains debug symbols, so that when a stack trace
containing e.g. BCL stack frames is sent to Microsoft, there is a way
to correlate that back to filename and line number source for further
investigation.

Xamarin.Android has not partaken in this mechanism, in large part
because we did most of our building on Linux and macOS machines which
generate `.mdb` debug symbols, while symbol servers require `.pdb`.

That constraint is now behind us; with 76be1fb1 and the use of
Mono 4.9 for builds, our builds now generate Portable `.pdb` files,
which *should* (hopefully) be usable on Microsoft's symbol servers.

However, to transmit them to Microsoft's servers, we need to actually
*retain* them. At present, they're not: because of `bundle*.zip` use,
the `mscorlib.dll` that we include in the final product may have been
built much earlier than the rest of the product, and we don't have any
other mechanism to retain the `mscorlib.pdb` file.

Let's change that: update `mono-runtimes.targets` so that the created
debug symbol files are included in the produced `bundle*.zip`, and
bump the corresponding `$(XABundleFileName)` version number. This will
ensure that we retain the debug symbols for our BCL assemblies,
allowing us to actually have something to upload.

[0]: https://msdn.microsoft.com/en-us/library/cc667410.aspx