[Xamarin.Android.Build.Tasks] Revert changes from f06bcbe2 (#526)
The commit f06bcbe2 incorrectly changed the way we calculate
$(AndroidIncludeDebugSymbols). That value should ONLY be
enabled if DebugType is Full or Embedded. I.e if we have a full
on debug build.

It turns out the source of the problem was the fact that msbuild
under OSX was completely ignoring the $(DebugType) property in
the csproj. This caused it to always default to 'portable' which
is NOT a full debug build.