[Xamarin.ProjectTools] Always ref Microsoft.Build.Engine
@borgdylan is reporting that xamarin-android isn't building for him
with a more recent mono/master build, because the
`Microsoft.Build.BuildEngine` namespace can't be resolved, because the
`Microsoft.Build.Engine.dll` assembly isn't being referenced.

`Microsoft.Build.Engine.dll` isn't being referenced because it's
Conditional so that it's only referenced on Windows.

I don't know why this is conditional to Windows, but
`Microsoft.Build.Engine.dll` exists *now*, so always use it.