[mono-runtimes] `cross-arm64.exe` must be a 64-bit app
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=54353

When using `mono` as an AOT cross-compiler, mono's "bitness" needs to
match the "bitness" of the target ABI. For example, to generate AOT
native code for 32-bit armeabi, the "host" `mono` needs to run in a
32-bit process. Similarly, to generate AOT native code for 64-bit
aarch64-v8a ("arm64-v8a"), the "host" `mono` needs to run in a 64-bit
process.

Unfortunately...that wasn't the case:

	$ file bin/Debug/lib/mandroid/cross-arm64.exe
	bin/Debug/lib/mandroid/cross-arm64.exe: PE32 executable (console) Intel 80386, for MS Windows

The result is that when attempting to generate AOT native libraries
for arm64 *on Windows*, it would fail:

	$ xbuild /p:Configuration=Release /p:AndroidSupportedAbis=arm64-v8a /p:AotAssemblies=True /t:SignAndroidPackage Project.csproj
	...
	[aot-compiler stderr] Can't cross-compile on 32-bit platforms to 64-bit architecture.

(It would work on macOS, as `bin/Debug/bin/cross-arm64` is a 64-bit
executable, as is required.)

The fix is to correct the `@(_MonoCrossRuntime)` value for
`cross-arm64-win` so that it used the
`$(MingwCommandPrefix64)`-related values, *not* the
`$(MingwCommandPrefix32)`-related values. This ensures that we use
compiler settings to generate 64-bit binaries, resolving the error:

	$ file bin/Debug/lib/mandroid/cross-arm64.exe
	bin/Debug/lib/mandroid/cross-arm64.exe: PE32+ executable (console) x86-64, for MS Windows