[Xamarin.Android.Build.Tests] Switch over to use the system Nuget.
We are starting to see Nuget packages only be available
from the v3 repositories. Our tests are using Nuget.Core v2.
But Nuget v3 does not have the same API as v2 if you want to
manually restore packages. In fact the api is "unstable" and it
not simple to use. What was a couple of lines of code now has to
be a mess of subclasses and other support files. The Nuget for
those have also exploded into a mass of 24+ packages.

So easiest solution is to use the installed version of Nuget to
restore the packages. This means it is (or should be) up to date
and will work with both v2 and v3.

I have tested this on windows and mac and it works ok.