Always set TargetFrameworkRootPath in props as well as FrameworkPathOverride
By doing this in props, we allow users to override the value if they choose to.

Also, if $(TargetFrameworkRootPath) isn't set by us, it's not set in the
project evaluation phase, so by the time we set $(FrameworkPathOverride)
it's still empty and wrong for VS2013/15.

The newly added Sdk.targets also allows us to add the global targets override
Xamarin.Sdk.targets to be consistent with the Xamarin.Sdk.props that we already
import. This allows IDEs to provide additional behavior as needed that is
cross-cutting across all Xamarin builds.