[Xamarin.Android.Build.Tasks] Set TargetFrameworkRootPath in props (#573)
Always set `$(TargetFrameworkRootPath)` in props as well as
`$(FrameworkPathOverride)`, as this allows users to override the
value if they choose to.

Also, if `$(TargetFrameworkRootPath)` isn't set by us, it's not set
in the project evaluation phase, so by the time we set
`$(FrameworkPathOverride)` it's still empty and wrong for
Visual Studio 2013/2015.

The newly added `Xamarin.Android.Sdk.targets` also allows us to add
the global targets override `Xamarin.Sdk.targets` to be consistent
with the `Xamarin.Sdk.props` file that we already import. This allows
IDEs to provide additional behavior as needed that is cross-cutting
across all Xamarin builds.