[build] $(ALL_FRAMEWORKS) must use spaces, not tabs
(From the "You've got to be kidding me!" department.)

Commit 571546a1 [broke][0] the [full xamarin-android build][1]:

	src/OpenTK-1.0/OpenTK.csproj: error : …/xamarin-android/src/OpenTK-1.0/OpenTK.csproj:
	…/xamarin-android/src/OpenTK-1.0/OpenTK.csproj could not import
	"$(OutputPath)\..\..\..\xbuild\Xamarin\Android\Xamarin.Android.CSharp.targets"

That's a bit of a head-scratcher. *Why* can't that file be imported?
It certainly exists!

Turns out™, it's an `xbuild` "bug": if `$(OutputPath)` doesn't exist,
then you can't use `$(OutputPath)` as a relative path:

	# v7.1 is the default framework version; "remove" it:
	$ mv bin/Debug/lib/xbuild-frameworks/MonoAndroid/v7.1/ bin/Debug/lib/xbuild-frameworks/MonoAndroid/__v7.1

	# OpenTK.csproj now can't build:
	$ xbuild src/OpenTK-1.0/OpenTK.csproj
	src/OpenTK-1.0/OpenTK.csproj: error : …/xamarin-android/src/OpenTK-1.0/OpenTK.csproj:
	…/xamarin-android/src/OpenTK-1.0/OpenTK.csproj could not import
	"$(OutputPath)\..\..\..\xbuild\Xamarin\Android\Xamarin.Android.CSharp.targets

`msbuild` works fine.

Thus, the `xbuild` error means that `$(OutputPath)` does not yet
exist. Which is odd because the failing `xbuild` invocation is part of
`make opentk-jcw`, which runs after `make framework-assemblies`, so
`$(OutputPath)` most certainly *should* exist!

...except it doesn't: commit 571546a1 adds a new {API=26, ID=0, v7.99.0}
framework "tuple", but Jenkins doesn't contain an e.g.
`bin/Debug/lib/xbuild-frameworks/MonoAndroid/v7.99.0` directory.

This can also be manually reproduced:

	$ make framework-assemblies CONFIGURATIONS=Debug API_LEVELS="25 26"

After execution, there likewise is no
`bin/Debug/lib/xbuild-frameworks/MonoAndroid/v7.99.0` directory.

The build breaks because `make framework-assemblies` doesn't build the
API-O `Mono.Android.dll` binding.

The API-O `Mono.Android.dll` binding isn't built because of...tabs.

Yes. Seriously. *Tabs*.

The `make framework-assemblies` target uses a shell pipeline to
determine the current framework version:

	CUR_VERSION=`echo "$(ALL_FRAMEWORKS)"|tr -s " "|cut -d " " -s -f $(a)`

where `$(a)` is the API level, from `$(API_LEVELS)`.

When `$(a)` is `26`, `$CUR_VERSION` is the empty string:

	$ echo "_ _ _ _ _ _ _ _ _ v2.3  _   _   _   _   v4.0.3  v4.1  v4.2 v4.3  v4.4  v4.4.87   v5.0  v5.1  v6.0  v7.0  v7.1	v7.99.0" | tr -s " "|cut -d " " -s -f 26
	# empty string

The failure is due to the interaction of:

 1. Commit 571546a1 uses a *tab* in `$(ALL_FRAMEWORKS)`, not a space.
    The tab should also be present in the above `echo` command,
    between "v7.1" and "v7.99.0".
 2. The above pipeline uses `cut -d " "`, which specifies *space* as
    the delimiter.

The result is an inconsistency: the tab is never replaced with a
space, and the `cut -d " " -s -f 26` looks for the 26th field in a
*space*-delimited line, but there is no 26th field; the *25th* field
has the value "v7.1<TAB>v7.99.0".

A proper fix would probably involve using `awk` instead of `cut -f`,
or replacing this mess with more MSBuild voodoo.

In the meantime, replace the <TAB> within `$(ALL_FRAMEWORKS)` with
spaces (vertical alignment is nice!), and "document" the that
`$(ALL_FRAMEWORKS)` and the other variables *must* use spaces.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/364/consoleText
[1]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/364/