[Xamarin.Android.Build.Tasks] Fix a stack overflow because LogError was calling itself.
Commit 049b2c8f introduced some default parameters for the LogError
method on the AsyncTask. This inadvertently causes one of the other
methods to start calling itself. This commit explicitly adds the
optional parameter names to make sure the correct method is being
called. This is because ALL LogError methods need to go through the

	protected void LogError (string code, string message, string file = null, int lineNumber = 0)

because it handles queuing the error and making sure that the data
gets marshalled back to the UI thread for logging. Not using this
method would cause UI lockups in VS.