[Xamarin.Android.Build.Tasks] Fix a stack overflow in LogError() (#576)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=55687
Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems?id=410362&fullScreen=false&_a=edit

Commit 049b2c8f introduced some default parameters for the `LogError()`
method on the `AsyncTask`. This inadvertently causes one of the other
methods to start calling itself. This commit explicitly adds the
optional parameter names to make sure the correct method is being
called. This is because *all* `LogError()` methods need to go through the

	protected void LogError (string code, string message, string file = null, int lineNumber = 0)

because it handles queuing the error and making sure that the data
gets marshalled back to the UI thread for logging. Not using this
method would cause UI lockups in VS.