[create-vsix] Allow overriding the created .vsix filename (#577)
We want to use the `make create-vsix` target to create the commercial
`Xamarin.Android.Sdk.vsix` file, but we want it to have a different
filename, matching commit + commit count information from the internal
`monodroid` repo, not the xamarin-android repo.

Uppdate `create-vsix.csproj` to accept a new `$(VsixPath)` property,
which is the path of the .vsix file to create.

Update the default `$(VsixPath)` filename to be
`Xamarin.Android.Sdk-OSS*`, so that if a customer downloads both a
Jenkins OSS build and a commercial .vsix build, it will be easier to
differentiate between them. (The `.vsix` package name information is
*not* changed, thus it is not possible to concurrently have both the
commercial and OSS builds installed in the same VS 2017 SKU.)

Update the `make create-vsix` target so that the output path and other
MSBuild properties can more easily be overridden:

	make create-vsix CONFIGURATIONS=Debug VSIX=`pwd`/MyNewPackage.vsix

The `$(VSIX)` **make**(1) variable sets the `$(VsixPath)` MSBuild
property.

*Note*: the current-working-directory for `make create-vsix` is
`build-tools/create-vsix`; consequently, this command:

	make create-vsix CONFIGURATIONS=Debug VSIX=MyNewPackage.vsix

will create `./build-tools/create-vsix/MyNewPackage.vsix`, *not*
`./MyNewPackage.vsix`.