[Xamarin.Android.Build.Tasks] Support multidex + ' ' in SDK path (#575)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=33052
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=38693

The google tooling around multidex is pretty broken on
windows. It does not handle paths with spaces and has
over quoting path issues. This is one reason why we are
shipping our own Proguard.

This commit completely removes the use of `mainDexClasses.bat`
and introduces code which will create the ` multidex.keep` by
making the required calls manually.