[Mono.Android] Fully support API-O Rebuilds
Commit eb08bb46 had the right idea, but didn't fix what it said:

	# claim
	$ make framework-assemblies API_LEVELS=26 CONFIGURATIONS=Debug
	$ make framework-assemblies API_LEVELS=26 CONFIGURATIONS=Debug
	# no errors!

In fact, the 2nd `make framework-assemblies` *still* produced 170582
errors; the test case was not in fact fixed.

What *was* fixed as building `Mono.Android.csproj` itself:

	$ xbuild src/Mono.Android/Mono.Android.csproj \
		/p:AndroidApiLevel=26 /p:AndroidPlatformId=O /p:AndroidFrameworkVersion=v7.99.0
	$ xbuild src/Mono.Android/Mono.Android.csproj \
		/p:AndroidApiLevel=26 /p:AndroidPlatformId=O /p:AndroidFrameworkVersion=v7.99.0
	# no errors

If `Mono.Android.csproj` was fixed, why was
`make framework-assemblies` still broken? Because the
`framework-assemblies` target *also* rebuilds
`src/Xamarin.Android.NUnitLite/Xamarin.Android.NUnitLite.csproj` --
which has a `@(ProjectReference)` on `Mono.Android.csproj --
and the `Xamarin.Android.NUnitLite.csproj` invocation did *not*
specify a `$(AndroidPlatformId)` override.

The result is a rebuild:

	$ make framework-assemblies API_LEVELS=26 CONFIGURATIONS=Debug
	$ make framework-assemblies API_LEVELS=26 CONFIGURATIONS=Debug
	...
	Target _GenerateApiDescription needs to be built as input file 'Profiles/api-26.xml.in' does not exist.
	...
	**BOOM**

It's looking for `api-26.xml.in` because `$(AndroidPlatformId)` isn't
overridden, so it defaults to using `$(AndroidApiLevel)`, which is 26.

A complete fix for `make framework-assemblies`, then, is to also
provide `$(AndroidPlatformId)` to the
`Xamarin.Android.NUniteLite.csproj` build command, thus preventing the
`_GenerateApiDescription` target from rebuilding.