[Xamarin.Android.Build.Tests] Migrate most of the remaining unit tests for msbuild.
We had a bunch of unit tests left over which need to be migrated.
This commit brings over most the those tests. Excluding Wear tests
since we don't install the wear api's by default, and tests to do
with FastDev.

We need a new BuildTest.OSS.cs file because we made the BuildTest
class partial. This allows us to pass in slightly different options
to the tests for the OSS repo. Mostly this is to do with the abi's
which are available. For example abi armeabi is NOT built by
default in this repo, but armeabi-v7a and x86 are.